move user views to users directory
